Moved the service logic out of the users controller
refs #12537
